refactor: Token hero animation components
